Rename Callbakcs to register
